requirements.txt
lxml==4.6.3
pandas==1.1.5
PyExecJS==1.5.1
requests==2.26.0
xlrd==1.2.0
